+++ landingzones/tests/test_tasks_taskflow.py
@@ -37,7 +37,7 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
             AppAlert.objects.filter(
                 app_plugin__name=APP_NAME,
                 alert_name=name,
-                user=self.landing_zone.user,
+                user=self.zone.user,
                 project=self.project,
             ).count(),
             count,
@@ -57,7 +57,7 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
     def setUp(self):
         super().setUp()
         self.task_kw = {
-            'landing_zone': self.landing_zone,
+            'landing_zone': self.zone,
             'flow_name': 'landing_zone_move',
             'status': lc.ZONE_STATUS_MOVED,
             'status_info': lc.DEFAULT_STATUS_INFO[lc.ZONE_STATUS_MOVED],
@@ -66,9 +66,9 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
 
     def test_execute(self):
         """Test SetLandingZoneStatusTask execute()"""
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self.assertEqual(
-            self.landing_zone.status_info,
+            self.zone.status_info,
             lc.DEFAULT_STATUS_INFO[lc.ZONE_STATUS_ACTIVE],
         )
         self._assert_owner_alert(0)
@@ -76,11 +76,11 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
         self.assertEqual(len(mail.outbox), 0)
 
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
+        self.zone.refresh_from_db()
 
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_MOVED)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_MOVED)
         self.assertEqual(
-            self.landing_zone.status_info,
+            self.zone.status_info,
             lc.DEFAULT_STATUS_INFO[lc.ZONE_STATUS_MOVED],
         )
         self._assert_owner_alert(1)
@@ -94,13 +94,13 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
     @override_settings(PROJECTROLES_SEND_EMAIL=False)
     def test_execute_disable_email(self):
         """Test execute() with email disabled"""
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_MOVED)
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_MOVED)
         self._assert_owner_alert(1)
         self._assert_member_alerts(1)
         self.assertEqual(len(mail.outbox), 0)
@@ -110,13 +110,13 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
         app_settings.set(
             APP_NAME, 'member_notify_move', False, project=self.project
         )
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_MOVED)
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_MOVED)
         self._assert_owner_alert(1)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 1)
@@ -128,15 +128,15 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
             APP_NAME,
             'notify_email_zone_status',
             False,
-            user=self.landing_zone.user,
+            user=self.zone.user,
         )
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_MOVED)
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_MOVED)
         self._assert_owner_alert(1)
         self._assert_member_alerts(1)
         self.assertEqual(len(mail.outbox), 1)
@@ -146,14 +146,14 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
 
     def test_execute_moved_no_files(self):
         """Test execute() with a busy status"""
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
         self.task_kw['extra_data'] = {'file_count': 0}
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_MOVED)
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_MOVED)
         # No alerts or emails should be set
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
@@ -161,7 +161,7 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
 
     def test_execute_busy(self):
         """Test execute() with busy status"""
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
@@ -170,10 +170,10 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
             lc.ZONE_STATUS_MOVING
         ]
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_MOVING)
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_MOVING)
         self.assertEqual(
-            self.landing_zone.status_info,
+            self.zone.status_info,
             lc.DEFAULT_STATUS_INFO[lc.ZONE_STATUS_MOVING],
         )
         # No alerts or emails should be set
@@ -183,7 +183,7 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
 
     def test_execute_failed(self):
         """Test execute() with FAILED status"""
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
@@ -192,10 +192,10 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
             lc.ZONE_STATUS_FAILED
         ]
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_FAILED)
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_FAILED)
         self.assertEqual(
-            self.landing_zone.status_info,
+            self.zone.status_info,
             lc.DEFAULT_STATUS_INFO[lc.ZONE_STATUS_FAILED],
         )
         self._assert_owner_alert(1)
@@ -208,7 +208,7 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
 
     def test_execute_validate(self):
         """Test execute() in validate mode"""
-        self.assertEqual(self.landing_zone.status, lc.ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.zone.status, lc.ZONE_STATUS_ACTIVE)
         self._assert_owner_alert(0)
         self._assert_member_alerts(0)
         self.assertEqual(len(mail.outbox), 0)
@@ -218,7 +218,7 @@ class TestSetLandingZoneStatusTask(ViewTestBase):
         ]
         self.task_kw['extra_data'] = {'validate_only': True}
         self._get_task().execute(**self.task_kw)
-        self.landing_zone.refresh_from_db()
+        self.zone.refresh_from_db()
         self.task_kw['status'] = lc.ZONE_STATUS_ACTIVE
         self.task_kw['status_info'] = [
             lc.DEFAULT_STATUS_INFO[lc.ZONE_STATUS_ACTIVE]